240722 demo edited

+++ demo/src/main/java/com/example/demo/model/MatchResultDto.java
@@ -3,7 +3,7 @@ package com.example.demo.model;
 public interface MatchResultDto {
 
     int getWhole_game();
-    int getSot_percent();
-    int getPoss_percent();
-    int getTac_percent();
+    double getSot_percent();
+    double getPoss_percent();
+    double getTac_percent();
 }
